pyproject.toml
[project]
name = "starter"
version = "0.1.0"
description = ""
authors = [
    {name = "Ahmed Elshinnawi",email = "ahmed.a.elshinnawi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11 , < 4"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-unfold (>=0.63.0,<0.64.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "django-cotton (>=2.1.3,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "django-modeltranslation (>=0.19.16,<0.20.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"


[tool.poetry.group.dev.dependencies]
django-environ = "^0.12.0"
django-rosetta = "^0.10.2"
django-debug-toolbar = "^6.0.0"

